Modularizacion de ServerPage.dart
